Infer gopath when debugging if program doesnt fall under current gopath
