+++ src/goPackages.ts
@@ -1,8 +1,8 @@
 import vscode = require('vscode');
 import cp = require('child_process');
 import path = require('path');
-import { getGoRuntimePath } from './goPath';
-import { isVendorSupported, getCurrentGoWorkspaceFromGOPATH, getToolsEnvVars } from './util';
+import { getGoRuntimePath, getCurrentGoWorkspaceFromGOPATH } from './goPath';
+import { isVendorSupported, getCurrentGoPath, getToolsEnvVars } from './util';
 
 let allPkgs = new Map<string, string>();
 let goListAllCompleted: boolean = false;
@@ -62,7 +62,7 @@ export function getImportablePackages(filePath: string): Promise<Map<string, str
 	return Promise.all([isVendorSupported(), goListAll()]).then(values => {
 		let isVendorSupported = values[0];
 		let currentFileDirPath = path.dirname(filePath);
-		let currentWorkspace = getCurrentGoWorkspaceFromGOPATH(currentFileDirPath);
+		let currentWorkspace = getCurrentGoWorkspaceFromGOPATH(getCurrentGoPath(), currentFileDirPath);
 		let pkgMap = new Map<string, string>();
 
 		allPkgs.forEach((pkgName, pkgPath) => {
